+++ src/main/java/doore/document/domain/repository/DocumentRepository.java
@@ -9,5 +9,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface DocumentRepository extends JpaRepository<Document, Long> {
     Page<Document> findAllByGroupTypeAndGroupId(DocumentGroupType groupType, Long groupId, Pageable pageable);
 
-    int countByGroupId(Long groupId);
+    Long countByGroupId(Long groupId);
 }
